services_generate-follow-up_1746474451.0/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-secret-manager
google-generativeai
pandas
pandas-gbq
